Enable support for Node.js v10
But this is only on a branch, as it requires pinning the version of the
openid-client module to an old version that will never be updated. This is
only useful for those customers that absolutely must run the service on a
system that absolutely cannot run Node.js v12 (CentOS/RHEL/Oracle 6).